fix: uint->ulong for some object ids

+++ Dalamud/Game/ClientState/Objects/ObjectTable.cs
@@ -60,7 +60,7 @@ public sealed partial class ObjectTable : IServiceType
     /// </summary>
     /// <param name="objectId">Object ID to find.</param>
     /// <returns>A game object or null.</returns>
-    public GameObject? SearchById(uint objectId)
+    public GameObject? SearchById(ulong objectId)
     {
         if (objectId is GameObject.InvalidGameObjectId or 0)
             return null;
